Better install scripts

+++ git/remove.sh
@@ -1,3 +1,8 @@
 #!/bin/bash
 
-rm "$HOME/.gitconfig"
+if [[ -f "$HOME/.gitconfig" ]]; then
+  echo "git: removing $HOME/.gitconfig [OK]"
+  rm "$HOME/.gitconfig"
+else
+  echo "git: $HOME/.gitconfig is already removed [OK]"
+fi
