+++ sublime/remove.sh
@@ -1,8 +1,26 @@
 #!/bin/bash
 source variables.sh
 
-rm "$package_path/$package_control"
-touch "$package_path/$package_control"
+# Package Control.sublime-settings
+if [[ -f "$package_path/$package_control" ]]; then
+  echo "sublime: removing $package_path/$package_control [OK]"
+  rm "$package_path/$package_control"
+  echo "sublime: creating empty $package_path/$package_control [OK]"
+  touch "$package_path/$package_control"
+else
+  echo "sublime: $package_path/$package_control is already removed [OK]"
+  echo "sublime: creating empty $package_path/$package_control [OK]"
+  touch "$package_path/$package_control"
+fi
 
-rm "$package_path/$preferences"
-touch "$package_path/$preferences"
+# Preferences.sublime-settings
+if [[ -f "$package_path/$preferences" ]]; then
+  echo "sublime: removing $package_path/$preferences [OK]"
+  rm "$package_path/$preferences"
+  echo "sublime: creating empty $package_path/$preferences [OK]"
+  touch "$package_path/$preferences"
+else
+  echo "sublime: $package_path/$preferences is already removed [OK]"
+  echo "sublime: creating empty $package_path/$preferences [OK]"
+  touch "$package_path/$preferences"
+fi
